fix(settings): read rate_limit_rps via Float64, not Int
rate_limit_rps is declared KindFloat, but the limiter read it as
settingsRuntime.Int(): json.Unmarshal of a fractional value like 2.5 into
int fails and returns 0, silently disabling the per-IP rate limit set
from the console. Read it via Float64 in both the boot SetRate and the
settings watcher callback.

Audited every settingsRuntime.Int/Float64/String/Bool/Duration consumer
against the Key* catalog kinds — this was the only mismatch (all Int
keys are read via Int/Duration, all String via String, all Bool via
Bool; the other Float key llm_temperature was already read via Float64).
The settings test now seeds a fractional 2.5 and asserts the Float64
read pins the regression.

+++ cmd/server/main.go
@@ -1861,14 +1861,16 @@ func httpHandler(ctx context.Context, cfg config.Config, log *slog.Logger, metri
 	// Live retune: pick up the runtime settings (0/0 = disabled); existing
 	// buckets converge within the limiter's sweep TTL. The settings watcher
 	// keeps the rate in sync with the admin console, so retuning
-	// rate_limit_rps / rate_limit_burst needs no restart.
+	// rate_limit_rps / rate_limit_burst needs no restart. rps is a KindFloat
+	// key — read via Float64, or a fractional value (e.g. 2.5) would unmarshal
+	// into int as 0 and silently disable the limiter.
 	limiter.SetRate(
-		float64(settingsRuntime.Int(settings.KeyRateLimitRPS)),
+		settingsRuntime.Float64(settings.KeyRateLimitRPS),
 		settingsRuntime.Int(settings.KeyRateLimitBurst),
 	)
 	settingsSync.Add(func() {
 		limiter.SetRate(
-			float64(settingsRuntime.Int(settings.KeyRateLimitRPS)),
+			settingsRuntime.Float64(settings.KeyRateLimitRPS),
 			settingsRuntime.Int(settings.KeyRateLimitBurst),
 		)
 	})
